Same for hotline and people.
