+++ frontend/src/people/PersonForm.js
@@ -8,10 +8,10 @@ import { AddressLookup, DateTimePicker, DropDown, TextInput } from '../component
 import { AuthContext } from "../accounts/AccountsReducer";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faArrowAltCircleLeft } from '@fortawesome/free-solid-svg-icons';
-import { STATE_OPTIONS } from '../constants'
+import { STATE_OPTIONS } from '../constants';
 
-// Form for creating new owner and reporter Person objects.
-export const PersonForm = (props) => {
+// Form for owner and reporter Person objects.
+const PersonForm = (props) => {
 
   const { state, dispatch } = useContext(AuthContext);
   const id = props.id;
@@ -186,10 +186,10 @@ export const PersonForm = (props) => {
                 navigate(state.prevLocation);
               }
               else if (isOwner) {
-                navigate('/hotline/owner/' + id);
+                navigate('/people/owner/' + id);
               }
               else {
-                navigate('/hotline/reporter/' + id);
+                navigate('/people/reporter/' + id);
               }
             })
             .catch(error => {
@@ -209,7 +209,7 @@ export const PersonForm = (props) => {
               }
               // If adding from an owner, redirect to the new Owner details.
               else if (owner_id) {
-                navigate('/hotline/owner/' + response.data.id);
+                navigate('/people/owner/' + response.data.id);
               }
               // If we have a reporter ID, redirect to create a new Animal with owner + reporter IDs.
               else if (reporter_id) {
@@ -384,134 +384,4 @@ export const PersonForm = (props) => {
   );
 };
 
-export const OwnerContactForm = ({id}) => {
-
-  const [queryParams] = useQueryParams();
-  var url;
-  var axios_method;
-
-  const {
-    owner = null,
-  } = queryParams;
-
-    const [data, setData] = useState({
-      owner_name: '',
-      owner_contact_time: '',
-      owner_contact_note: '',
-      owner: owner,
-    })
-
-    useEffect(() => {
-      let source = axios.CancelToken.source();
-      if (id) {
-        // Get OwnerContact if editing existing
-        const fetchOwnerContact = async () => {
-          // Fetch Owner Contact data.
-          await axios.get('/people/api/ownercontact/' + id + '/', {
-            cancelToken: source.token,
-          })
-          .then(response => {
-            setData(response.data);
-            console.log(response.data)
-          })
-          .catch(error => {
-            console.log(error.response);
-          });
-        };
-        fetchOwnerContact();
-      }
-      else {
-        // Get Owner details for new OwnerContact
-        const fetchOwner = async () => {
-        // Fetch Owner Data.
-        axios.get('/people/api/person/' + owner +'/', {
-          cancelToken: source.token,
-        })
-          .then(response => {
-            setData(prevState => ({ ...prevState, ["owner_name"]:response.data['first_name'] + ' ' + response.data['last_name'] }));
-          })
-          .catch(error => {
-            console.log(error.response);
-          });
-        }
-        fetchOwner();
-      };
-      return () => {
-        source.cancel();
-      };
-    }, [id]);
-
-    return (
-        <Formik
-          initialValues={data}
-          enableReinitialize={true}
-          validationSchema={Yup.object({
-            owner_contact_time: Yup.date().required(),
-            owner_contact_note: Yup.string().required(),
-          })}
-          onSubmit={(values, { setSubmitting }) => {
-            setTimeout(() => {
-              if (id) {
-                url = url + values.id
-                axios_method = axios.patch
-              }
-              else {
-                url = '/people/api/ownercontact/'
-                axios_method = axios.post
-              }
-              axios_method(url, values)
-              .then(
-                  navigate('/hotline/owner/' + values.owner)
-              )
-              .catch(error => {
-                console.log(error.response);
-              });
-            setSubmitting(false);
-            }, 500);
-          }}
-        >
-        {form => (
-          <Card border="secondary" className="mt-5">
-          <Card.Header as="h5" className="pl-3"><span style={{cursor:'pointer'}} onClick={() => window.history.back()} className="mr-3"><FontAwesomeIcon icon={faArrowAltCircleLeft} size="lg" inverse /></span>{ data.owner_name } - {!id ? "New" : "Update"} Owner Contact</Card.Header>
-          <Card.Body>
-          <BootstrapForm>
-              <FormGroup>
-                <Row>
-                  <Col >
-                  <DateTimePicker
-                    label="Owner Contact Time"
-                    name="owner_contact_time"
-                    id="owner_contact_time"
-                    xs="7"
-                    clearable={false}
-                    data-enable-time={true}
-                    onChange={(date, dateStr) => {
-                      form.setFieldValue("owner_contact_time", dateStr)
-                    }}
-                    value={form.values.owner_contact_time||null}
-                  />
-                  </Col>
-                </Row>
-                <Row className="mt-3">
-                  <Col >
-                    <TextInput
-                      as="textarea"
-                      label="Owner Contact Note"
-                      name="owner_contact_note"
-                      id="owner_contact_note"
-                      xs="7"
-                      rows={5}
-                    />
-                  </Col>
-                </Row>
-              </FormGroup>
-          </BootstrapForm>
-          </Card.Body>
-          <ButtonGroup>
-            <Button type="button" className="btn btn-primary" onClick={() => {form.submitForm()}}>Save</Button>
-          </ButtonGroup>
-          </Card>
-          )}
-        </Formik>
-    );
-};
+export default PersonForm;
